requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyautogui>=0.9.54
mss>=9.0.1
pytesseract>=0.3.10
PyQt6>=6.6.0
scipy>=1.11.0
